+++ index.ts
@@ -1,7 +1,7 @@
-import AutoIndex from "./src/AutoIndex.vue"
-import HighText from "./src/HighText.vue"
-import InputFile from "./src/InputFile.vue"
-import ShowValue from "./src/ShowValue.vue"
+import AutoIndex from "./src/data/AutoIndex.vue"
+import HighText from "./src/data/HighText.vue"
+import InputFile from "./src/data/InputFile.vue"
+import ShowValue from "./src/data/ShowValue.vue"
 
 export {
   AutoIndex,