+++ src/mod/CommaNumber.vue
@@ -1,0 +1,57 @@
+
+<template>
+  <span v-if="!format" class="comma-number" :class="currentClass" >{{ currentData }}</span>
+  <span v-else class="comma-number" :class="currentClass" >{{ format(currentData) }}</span>
+</template>
+
+<script lang="ts">
+import { defineComponent, PropType } from 'vue'
+import config from '../../config'
+
+export type formatFunctionType = (n: number | string) => (number | string)
+
+export default defineComponent({
+  name: 'CommaNumber',
+  props: {
+    data: {
+      type: Number,
+      required: true
+    },
+    color: {
+      type: String,
+      required: false
+    },
+    decimal: {
+      type: Boolean,
+      required: false,
+      default: false
+    },
+    comma: {
+      type: String,
+      required: false,
+      default: ','
+    },
+    format: {
+      type: Function as PropType<formatFunctionType>,
+      required: false,
+      default: undefined
+    }
+  },
+  data() {
+    return {
+    }
+  },
+  computed: {
+    currentClass() {
+      if (this.color) {
+        return 'local-main-color-' + this.color
+      } else {
+        return ''
+      }
+    },
+    currentData() {
+      return config.parseNumberByComma(this.data, this.comma, this.decimal) as string
+    }
+  }
+})
+</script>