+++ history.md
@@ -1,4 +1,5 @@
 ### ToDo
+- 简化TableView，相同的参数实现基础的table，主要目标是优化性能
 
 ### 4.0.3/4/5/6/7/8
 - 升级依赖，适配AttributeValue